Working on mimicking the Kohana class / file system structure
