+++ classes/kacela/collection/arr.php
@@ -1,0 +1,3 @@
+<?php
+
+class Kacela_Collection_Arr extends Kohana_Kacela_Collection_Arr {}
