+++ classes/kacela/collection/statement.php
@@ -1,0 +1,3 @@
+<?php
+
+class Kacela_Collection_Statement extends Kohana_Kacela_Collection_Statement {}
