--- classes/kacela/datasource/adapter/adapter.php
@@ -1,17 +1,0 @@
-<?php
-
-/**
- * Description of adapter
- *
- * @author noah
- * @date $(date)
- */
-
-namespace Kacela\DataSource\Adapter;
-
-use Gacela\DataSource\Adapter as A;
-
-abstract class Adapter extends A\Adapter
-{
-
-}
