+++ classes/kacela/datasource/adapter/mysql.php
@@ -1,0 +1,3 @@
+<?php
+
+class Kacela_DataSource_Adapter_Mysql extends Kohana_Kacela_DataSource_Adapter_Mysql {}
