+++ classes/kacela/datasource/adapter/salesforce.php
@@ -1,0 +1,3 @@
+<?php
+
+class Kacela_DataSource_Adapter_Salesforce extends Kohana_Kacela_DataSource_Adapter_Salesforce {}
