--- classes/kacela/datasource/datasource.php
@@ -1,17 +1,0 @@
-<?php
-/**
- * @author noahg
- * @date 6/17/11
- * @brief
- *
- */
-
-namespace Kacela\DataSource;
-
-use Gacela\DataSource as D;
-
-abstract class DataSource extends D\DataSource
-{
-
-
-}
