+++ classes/kacela/datasource/query/soql.php
@@ -1,0 +1,5 @@
+<?php
+
+class Kacela_DataSource_Query_Soql extends Kohana_Kacela_DataSource_Query_Soql {
+
+}
