+++ classes/kacela/field/bool.php
@@ -1,0 +1,3 @@
+<?php
+
+class Kacela_Field_Bool extends Kohana_Kacela_Field_Field {}
