+++ classes/kacela/field/date.php
@@ -1,0 +1,3 @@
+<?php
+
+class Kacela_Field_Date extends Kohana_Kacela_Field_Field {}
