+++ classes/kacela/field/enum.php
@@ -1,0 +1,3 @@
+<?php
+
+class Kacela_Field_Enum extends Kohana_Kacela_Field_Field {}
