+++ classes/kacela/field/field.php
@@ -1,0 +1,3 @@
+<?php
+
+class Kacela_Field_Field extends Kohana_Kacela_Field_Field {}
