+++ classes/kacela/field/float.php
@@ -1,0 +1,3 @@
+<?php
+
+class Kacela_Field_Float extends Kohana_Kacela_Field_Field {}
