+++ classes/kacela/field/int.php
@@ -1,0 +1,3 @@
+<?php
+
+class Kacela_Field_Int extends Kohana_Kacela_Field_Field {}
