+++ classes/kacela/field/string.php
@@ -1,0 +1,3 @@
+<?php
+
+class Kacela_Field_String extends Kohana_Kacela_Field_Field {}
