--- classes/kacela/model/imodel.php
@@ -1,16 +1,0 @@
-<?php
-/** 
- * @author noahg
- * @date 6/16/11
- * @brief
- * 
- */
-
-namespace Kacela\Model;
-
-use Gacela\Model as M;
-
-interface iModel extends M\iModel
-{
-
-}
