+++ classes/kacela/profiler.php
@@ -6,10 +6,4 @@
  * @author noah
  * @date $(date)
  */
-class Kacela_Profiler extends Kohana_Profiler
-{
-	public static function set_name($token, $name)
-	{
-		Profiler::$_marks[$token]['name'] = $name;
-	}
-}
+class Kacela_Profiler extends Kohana_Kacela_Profiler {}
