+++ classes/kohana/kacela/collection/arr.php
@@ -1,0 +1,8 @@
+<?php
+
+use Gacela\Collection as C;
+
+class Kohana_Kacela_Collection_Arr extends C\Arr
+{
+
+}
