+++ classes/kohana/kacela/collection/statement.php
@@ -1,0 +1,8 @@
+<?php
+
+use Gacela\Collection as C;
+
+class Kohana_Kacela_Collection_Statement extends C\Statement
+{
+
+}
