+++ classes/kohana/kacela/datasource/adapter/mysql.php
@@ -1,0 +1,8 @@
+<?php
+
+use Gacela\DataSource\Adapter as A;
+
+class Kohana_Kacela_DataSource_Adapter_Mysql extends A\Mysql
+{
+
+}
