+++ classes/kohana/kacela/datasource/adapter/salesforce.php
@@ -1,0 +1,8 @@
+<?php
+
+use Gacela\DataSource\Adapter as A;
+
+class Kohana_Kacela_DataSource_Adapter_Salesforce extends A\Salesforce
+{
+
+}
