+++ classes/kohana/kacela/datasource/query/soql.php
@@ -1,0 +1,8 @@
+<?php
+
+use Gacela\DataSource\Query as Q;
+
+class Kohana_Kacela_DataSource_Query_Soql extends Q\Soql
+{
+
+}
