+++ classes/kohana/kacela/field/bool.php
@@ -1,0 +1,8 @@
+<?php
+
+use Gacela\Field as F;
+
+class Kohana_Kacela_Field_Bool extends F\Bool
+{
+
+}
