+++ classes/kohana/kacela/field/int.php
@@ -1,0 +1,8 @@
+<?php
+
+use Gacela\Field as F;
+
+class Kohana_Kacela_Field_Int extends F\Int
+{
+
+}
