+++ classes/kohana/kacela/field/string.php
@@ -1,0 +1,8 @@
+<?php
+
+use Gacela\Field as F;
+
+class Kohana_Kacela_Field_String extends F\String
+{
+
+}
